+++ modules/platform-sdl2/mqueue.cpp
@@ -223,7 +223,7 @@ imqd_t implementation::open(const char *name, int max_messages) {
 
 int implementation::close(imqd_t qd) {
     SDL_LockMutex(queues_mtx);
-    if (qd >= queues.size()) {
+    if (qd >= (int)queues.size()) {
         SDL_UnlockMutex(queues_mtx);
         return IMQ_BADF;
     }
@@ -257,7 +257,7 @@ int implementation::unlink(const char *name) {
 
 the_queue *implementation::get_queue(imqd_t qd) {
     SDL_LockMutex(queues_mtx);
-    if (qd >= queues.size()) {
+    if (qd >= (int)queues.size()) {
         SDL_UnlockMutex(queues_mtx);
         return NULL;
     }
